expose levelizeobserver (#433)
